confirm before transport deletion

+++ deltachat-ios/Controller/Settings/Transports/TransportListViewController.swift
@@ -63,12 +63,19 @@ class TransportListViewController: UITableViewController {
 
     private func deleteTransport(at indexPath: IndexPath) {
         guard let transport = transports.get(at: indexPath.row) else { return }
-        do {
-            _ = try self.dcContext.deleteTransport(addr: transport.addr)
-        } catch {
-            logAndAlert(error: error.localizedDescription)
-        }
-        reloadTransports()
+
+        let alert = UIAlertController(title: String.localized(stringID: "confirm_remove_transport", parameter: transport.addr), message: nil, preferredStyle: .safeActionSheet)
+        alert.addAction(UIAlertAction(title: String.localized("remove_transport"), style: .destructive, handler: { [weak self] _ in
+            guard let self else { return }
+            do {
+                _ = try self.dcContext.deleteTransport(addr: transport.addr)
+            } catch {
+                logAndAlert(error: error.localizedDescription)
+            }
+            reloadTransports()
+        }))
+        alert.addAction(UIAlertAction(title: String.localized("cancel"), style: .cancel))
+        present(alert, animated: true, completion: nil)
     }
 }
